source_base/AutoDev_IQ_BE/requirements.txt
fastapi
uvicorn
GitPython
langchain
langchain-community
langchain-core
langchain-ollama
matplotlib
networkx
javalang
chromadb
playwright
deepdiff
BeautifulSoup4
cssutils
